Address comments, move base attrs to subclasses
